teamsbot/requirements.txt
microsoft-agents-hosting-aiohttp
microsoft-agents-hosting-teams
microsoft-agents-authentication-msal
cryptography